[feat/#188] :recycle: Refactor: data 파일 내 코드 리팩토링 및 병합
