+++ model/.gitignore
@@ -2,8 +2,6 @@
 config.yaml
 dummyData
 __pycache__
-gradio_chached_examples
 faiss_index
 index.*
 log
-*_security.py
