+++ model/app/data/loader.py
@@ -24,6 +24,9 @@ class S3Loader:
             return None
 
 
+#############################################################################################
+
+
 class MinioLoader:
     def __init__(self, access_credential, bucket_name):
         self.bucket_name = bucket_name
